server/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
